FUTURE_SYSTEM/requirements.txt
gradio>=5.0.0
google-genai>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
unsloth
unsloth-zoo
modelscope
peft
accelerate
bitsandbytes
huggingface_hub
transformers>=4.51.3,<=4.57.6
